LoyalLight/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.3.7
sentence-transformers==2.2.2
PyPDF2==3.0.1
numpy==1.24.3
scikit-learn==1.3.2
slowapi==0.1.9
redis==5.0.1
aiofiles==23.2.1
Pillow==10.1.0

